:zap: Fix: 질문 삭제시시 작성자와 현재 로그인된 사용자 비교
Former-commit-id: 64013a8c3f7e061ded1f6059d18b8bb35b77e94e

+++ backend/src/main/java/com/ssafy/board/controller/BoardController.java
@@ -70,11 +70,8 @@ public class BoardController {
 
 			// 질문을 조회하여 작성자 정보 가져오기
 			BoardQuestionDto existingQuestion = boardService.getQuestionById(id);
-			System.out.println(authentication.getPrincipal().equals(existingQuestion.getAuthor()));
 
 			// 작성자와 현재 로그인된 사용자 비교
-			System.out.println(authentication.getAuthorities());
-
 			if (!existingQuestion.getAuthor().equals(authentication.getPrincipal())
 				&& !authentication.getAuthorities().contains("ROLE_admin")) {
 				return ResponseEntity.status(HttpStatus.FORBIDDEN).body("작성자가 아니므로 수정할 수 없습니다.");
@@ -91,8 +88,17 @@ public class BoardController {
 	// 질문 삭제
 	@DeleteMapping("/questions/{id}")
 	public ResponseEntity<String> deleteQuestion(@PathVariable
-	int id) {
+	int id, Authentication authentication) {
 		try {
+			// 질문을 조회하여 작성자 정보 가져오기
+			BoardQuestionDto existingQuestion = boardService.getQuestionById(id);
+
+			// 작성자와 현재 로그인된 사용자 비교
+			if (!existingQuestion.getAuthor().equals(authentication.getPrincipal())
+				&& !authentication.getAuthorities().contains("ROLE_admin")) {
+				return ResponseEntity.status(HttpStatus.FORBIDDEN).body("작성자가 아니므로 수정할 수 없습니다.");
+			}
+
 			boardService.deleteQuestion(id);
 			return ResponseEntity.ok("질문이 삭제되었습니다.");
 		} catch (Exception e) {
